test: Test mesh status of every alertmanager instance
